requirements.txt
fastapi
uvicorn
requests
python-dotenv
duckduckgo-search
aiofiles
jinja2
